move card file to component
